refactor: KanBanItem 과 KanBanItemInput 분리 [#87]
- KanBanItem 은 드래깅 하는 이벤트 관리할 예정이라 무거워질 것 같아서 분리
