+++ client/src/components/KanbanItem/index.tsx
@@ -2,23 +2,11 @@ import React, { useContext, useState } from 'react';
 import Styled from '@/components/KanbanItem/style';
 import { KanbanItemType } from '@/types/story';
 import { KanbanModalContext } from '@/components/KanbanModal';
-import { useStoryDispatch } from '@/lib/hooks/useContextHooks';
-import { useInput } from '@/lib/hooks';
-import { updateStoryWithName } from '@/lib/api/story';
+import { KanbanInput } from '@/components';
 
 const KanbanItem = ({ story, handleDragStart, handleDragEnter }: KanbanItemType) => {
   const modalConsumer = useContext(KanbanModalContext);
-  const dispatchStory = useStoryDispatch();
-  const { key, value, onChange } = useInput('');
   const [isDragEnter, setDragEnter] = useState(false);
-
-  const useUpdateStoryName = () => {
-    dispatchStory({
-      type: 'UPDATE_STORY',
-      story: { status: 'TODO', id: key, order: story.order, name: value },
-    });
-    updateStoryWithName({ status: 'TODO', id: key, order: story.order, name: value });
-  };
 
   return (
     <Styled.KanBanItem
@@ -36,13 +24,7 @@ const KanbanItem = ({ story, handleDragStart, handleDragEnter }: KanbanItemType)
       isDragEnter={isDragEnter}
       draggable
     >
-      <input
-        type="text"
-        placeholder={story.name ? story.name : 'type a todo...'}
-        data-key={story.id}
-        onChange={onChange}
-        onBlur={useUpdateStoryName}
-      />
+      <KanbanInput story={story} />
       <Styled.CancelIcon
         onClick={() => {
           modalConsumer?.setShowModal(true);
